Feature/prepare windows release for hopsy2

See merge request IBG-1/ModSim/hops!28

+++ src/hops/MarkovChain/Tuning/ExpectedSquaredJumpDistanceTuner.cpp
@@ -8,7 +8,7 @@ bool hops::ExpectedSquaredJumpDistanceTuner::tune(
         ExpectedSquaredJumpDistanceTuner::param_type& parameters,
         Eigen::MatrixXd& data) {
     auto target = ExpectedSquaredJumpDistanceTarget{markovChains, 
-                                                    parameters.iterationsToTestStepSize, 
+                                                    static_cast<unsigned long>(parameters.iterationsToTestStepSize), 
                                                     /*lags=*/{1},
                                                     parameters.considerTimeCost,
                                                     parameters.estimateCovariance};
